I04_1-156: renamed file names to match class name
