+++ dls_barcode/scan/scan_result.py
